+++ src/Piece/Bishop.java
@@ -1,0 +1,11 @@
+package Piece;
+
+public class Bishop extends Piece{
+	public Bishop(Boolean state) {
+		super ("Fou", state);
+	}
+	
+	public void play() {
+		
+	}
+}
