+++ src/Piece/EmptyPiece.java
@@ -1,0 +1,10 @@
+package Piece;
+public class EmptyPiece extends Piece{
+	public EmptyPiece() {
+		super(" ", true);
+	}
+	
+	public void play () {
+		
+	}
+}
