+++ src/Piece/King.java
@@ -1,0 +1,11 @@
+package Piece;
+
+public class King extends Piece{
+	public King(Boolean state) {
+		super("Roi", state);
+	}
+	
+	public void play() {
+		
+	}
+}
