+++ src/Piece/Knight.java
@@ -1,0 +1,11 @@
+package Piece;
+
+public class Knight extends Piece{
+	public Knight(Boolean state) {
+		super("Cavalier", state);
+	}
+	
+	public void play() {
+		
+	}
+}
