+++ src/Piece/Pawn.java
@@ -1,7 +1,7 @@
-package Pions;
+package Piece;
 public class Pawn extends Piece {
 	public Pawn (Boolean state) {
-		super("Pawn", state);
+		super("Piece", state);
 	}
 	
 	public void play() {
