+++ src/Piece/Queen.java
@@ -1,0 +1,11 @@
+package Piece;
+
+public class Queen extends Piece{
+	public Queen(Boolean state) {
+		super("Queen",state);
+	}
+	
+	public void play() {
+		
+	}
+}
