+++ src/Piece/readme.md
@@ -1,0 +1,9 @@
+Traduction des pièces 
+* Piece : pièce
+    * EmptyPiece : case vide
+    * Pawn: Pion
+    * Queen : Reine
+    * King : Roi
+    * Bishop : Fou
+    * Tower : Tour
+    * Knight : Cavalier
