--- src/Pions/Cavalier.java
@@ -1,11 +1,0 @@
-package Pions;
-
-public class Cavalier extends Piece{
-	public Cavalier (Boolean state) {
-		super("Cavalier", state);
-	}
-	
-	public void play() {
-		
-	}
-}
