--- src/Pions/Fou.java
@@ -1,11 +1,0 @@
-package Pions;
-
-public class Fou extends Piece{
-	public Fou (Boolean state) {
-		super ("Fou", state);
-	}
-	
-	public void play() {
-		
-	}
-}
