--- src/Pions/Reine.java
@@ -1,11 +1,0 @@
-package Pions;
-
-public class Reine extends Piece{
-	public Reine (Boolean state) {
-		super("Queen",state);
-	}
-	
-	public void play() {
-		
-	}
-}
