--- src/Pions/Roi.java
@@ -1,11 +1,0 @@
-package Pions;
-
-public class Roi extends Piece{
-	public Roi (Boolean state) {
-		super("Roi", state);
-	}
-	
-	public void play() {
-		
-	}
-}
