--- src/Pions/VoidPiece.java
@@ -1,10 +1,0 @@
-package Pions;
-public class VoidPiece extends Piece{
-	public VoidPiece () {
-		super(" ", true);
-	}
-	
-	public void play () {
-		
-	}
-}
